Move LSTM lifecycle coverage into RNN tests
